+++ src/WidgetSpaceshipStatus.cc
@@ -9,13 +9,13 @@ WidgetSpaceshipStatus::WidgetSpaceshipStatus(Widget * parent, const glm::ivec2 &
   : Widget(parent, position, size),
     _spaceship(nullptr)
 {
-  _meter = new MeshProgressBar(GetSize().x, GetSize().y, glm::vec3(0, 1, 0));
+  _meter = new MeshProgressBar(static_cast<float>(GetSize().x), static_cast<float>(GetSize().y), glm::vec3(0, 1, 0));
 }
 
 
 void WidgetSpaceshipStatus::Draw() const
 {
-  _meter->SetValue(glm::clamp(_spaceship->GetHealth() / 100.0, 0.0, 1.0));
+  _meter->SetValue(static_cast<float>(glm::clamp(_spaceship->GetHealth() / 100.0, 0.0, 1.0)));
   _meter->Draw(GetModel(), GetView(), GetProjection(), GetMVP());
   
   Widget::Draw();
